set LANG and LC_TYPE env variables to UTF-8 in zsh for tmux
The tmux from nixpkgs used with this home-manager configuration running
on MacOS had the issue that the powerline triangles were rendered as
underline.

This sets these environment variables in the shell profile so that they
are surely available the moment the user starts tmux.

Tmux handles UTF-8 correctly when the environment variables LANG/LC_TYPE
contain UTF-8.  (Alternatively tmux can be started with -u as well)