include ncurses from nixpkgs for the alacritty term type
Now because the `alacritty` terminal type is not the default terminfo
database on MacOS, the terminfo database from nixpkgs is now used.

The TERMINFO environment variable is automatically set to the
in the zprofile initialization script unless we are on NixOS
(which does this already correctly).

Now alacritty can finally announce the correct terminal type. This let's
tmux hopefully do the right things.