zsh: fix bell on prompt for tmux integration
The previous configuration had some problems when using zsh-suggestions
and zsh completions at the same time. In this case the first character
entered into the command line would be printed twice.

But this way is nicer to output the bell and it doesn't interfere with
redrawing the prompt/zsh-widgets after completion.