3mon: adjust monitor configuration to work fine with Hi-DPI display
This reconfigures the `3mon narrow` configuration mainly for now for my
new monitor setup with a Hi-DPI screen.

If we mix Hi-DPI screens with low-DPI screens we have a range of problems

* on high-dpi screens apps may look tiny
* on high-dpi screens apps may look blurry
* on low-dpi screens apps are way to big when scaled to high-dpi
* most apps can't change scaling when switching montior (or what if a
window laps over two monitors?)

But the solution is not to down-scale the hi-dpi screen (else it would
just look as if it had a lower resolution), but to up-scale all the
lower dpi screens to the same level.

The important part is the --scale-from option here. This specifies the
resolution the X-Applications can see. X will then scale it down
to the resolution specified via --mode.

The --right-of/--below-of <output> options could not be used for positioning here anymore as
the positioning is based upon the non scaled values and wrong.

This way it looks to all the applications that there are only hi-dpi
screens. That means I can set in the XFCE settings a higher dpi (144).
And tell applications like firefox to upscale their content.