tmux: remove -r from keybindings i often trigger on accident
It happens for example what I switch a pane and then press enter to
execute a command already entered.  With the previous keybinding this
resulted in the change of the pane layout.  This was especially
annoying when a lot of panes where setup.

Also I noticed what I sometimes accidentally swap more panes than I
want and I assume it's because of the -r bindings as well here in this
case.